Refactor: dto directory구조 변경
