+++ src/controllers/admin-controller.js
@@ -1,17 +1,15 @@
-const AdminDTO = require('../dtos/admin-dto');
+const AdminDTO = require('../dtos/admin-dto/admin-dto');
 const adminService = require('../services/admin-service');
 
-class AdminController {
-  static async loginAdmin(req, res) {
-    console.log(req.body);
+const loginAdmin = async (req, res) => {
+  const adminDto = new AdminDTO(req.body);
 
-    const adminDto = new AdminDTO(req.body);
-    console.log(adminDto.phoneNumber);
-    console.log(adminDto.password);
-
-    const token = await adminService.loginAdmin(adminDto);
-    res.status(200).json(token);
+  try {
+      const token = await adminService.loginAdmin(adminDto);
+      res.status(200).json(token);
+  } catch (error) {
+      res.status(401).json({ error: error.message });
   }
-}
+};
 
-module.exports = AdminController;
+module.exports = {loginAdmin};
