+++ src/models/certificate.js
@@ -1,0 +1,7 @@
+class Certificate {
+    constructor(path) {
+      this.path = path;
+    }
+}
+
+module.exports = Certificate;
